Move docs logic to controllers, limit passing Components service
