+++ src/main/kotlin/com/freya02/bot/commands/slash/docs/AnyFieldCommand.kt
@@ -1,7 +1,7 @@
 package com.freya02.bot.commands.slash.docs
 
 import com.freya02.bot.commands.slash.docs.CommonDocsHandlers.Companion.ANY_FIELD_NAME_AUTOCOMPLETE_NAME
-import com.freya02.bot.commands.slash.docs.CommonDocsHandlers.Companion.handleFieldDocs
+import com.freya02.bot.commands.slash.docs.controllers.SlashDocsController
 import com.freya02.bot.docs.DocIndexMap
 import com.freya02.bot.docs.index.DocSuggestion.Companion.mapToSuggestions
 import com.freya02.botcommands.api.annotations.CommandMarker
@@ -9,12 +9,11 @@ import com.freya02.botcommands.api.commands.application.ApplicationCommand
 import com.freya02.botcommands.api.commands.application.GlobalApplicationCommandManager
 import com.freya02.botcommands.api.commands.application.annotations.AppDeclaration
 import com.freya02.botcommands.api.commands.application.slash.GuildSlashEvent
-import com.freya02.botcommands.api.components.Components
 import com.freya02.docs.DocSourceType
 import dev.minn.jda.ktx.messages.reply_
 
 @CommandMarker
-class AnyFieldCommand(private val docIndexMap: DocIndexMap, private val components: Components) : ApplicationCommand() {
+class AnyFieldCommand(private val docIndexMap: DocIndexMap, private val slashDocsController: SlashDocsController) : ApplicationCommand() {
     @AppDeclaration
     fun declare(manager: GlobalApplicationCommandManager) {
         manager.slashCommand("anyfield") {
@@ -49,7 +48,7 @@ class AnyFieldCommand(private val docIndexMap: DocIndexMap, private val componen
         }
 
         val docIndex = docIndexMap[sourceType]!!
-        handleFieldDocs(event, split[0], split[1], docIndex, components) {
+        slashDocsController.handleFieldDocs(event, split[0], split[1], docIndex) {
             return@handleFieldDocs anyFieldNameAutocomplete(docIndex, fullSignature, 100).mapToSuggestions()
         }
     }
