+++ src/main/kotlin/com/freya02/bot/commands/slash/docs/SlashResolve.kt
@@ -1,18 +1,18 @@
 package com.freya02.bot.commands.slash.docs
 
 import com.freya02.bot.commands.slash.docs.CommonDocsHandlers.Companion.transformResolveChain
+import com.freya02.bot.commands.slash.docs.controllers.SlashDocsController
 import com.freya02.bot.docs.DocIndexMap
 import com.freya02.botcommands.api.annotations.CommandMarker
 import com.freya02.botcommands.api.commands.application.ApplicationCommand
 import com.freya02.botcommands.api.commands.application.GlobalApplicationCommandManager
 import com.freya02.botcommands.api.commands.application.annotations.AppDeclaration
 import com.freya02.botcommands.api.commands.application.slash.GuildSlashEvent
-import com.freya02.botcommands.api.components.Components
 import com.freya02.docs.DocSourceType
 import dev.minn.jda.ktx.messages.reply_
 
 @CommandMarker
-class SlashResolve(private val docIndexMap: DocIndexMap, private val components: Components) : ApplicationCommand() {
+class SlashResolve(private val docIndexMap: DocIndexMap, private val slashDocsController: SlashDocsController) : ApplicationCommand() {
     @AppDeclaration
     fun declare(manager: GlobalApplicationCommandManager) {
         manager.slashCommand("resolve") {
@@ -47,7 +47,7 @@ class SlashResolve(private val docIndexMap: DocIndexMap, private val components:
             event.reply_("Could not find documentation for `$chain`", ephemeral = true).queue()
             return
         }
-        CommonDocsHandlers.sendClass(event, false, doc, components)
+        slashDocsController.sendClass(event, false, doc)
     }
 
     companion object {
