+++ src/main/kotlin/com/freya02/bot/commands/text/TextDocs.kt
@@ -1,8 +1,8 @@
 package com.freya02.bot.commands.text
 
-import com.freya02.bot.commands.slash.docs.CommonDocsHandlers
+import com.freya02.bot.commands.controllers.CommonDocsController
+import com.freya02.bot.commands.text.docs.controllers.TextDocsController
 import com.freya02.bot.docs.DocIndexMap
-import com.freya02.bot.docs.index.DocIndex
 import com.freya02.bot.docs.index.DocSuggestion
 import com.freya02.bot.docs.index.DocSuggestion.Companion.mapToSuggestions
 import com.freya02.bot.docs.index.DocType
@@ -11,17 +11,15 @@ import com.freya02.botcommands.api.commands.prefixed.BaseCommandEvent
 import com.freya02.botcommands.api.commands.prefixed.TextCommand
 import com.freya02.botcommands.api.commands.prefixed.annotations.JDATextCommand
 import com.freya02.botcommands.api.commands.prefixed.annotations.TextOption
-import com.freya02.botcommands.api.components.Components
 import com.freya02.docs.DocSourceType
-import dev.minn.jda.ktx.messages.reply_
-import net.dv8tion.jda.api.exceptions.ErrorHandler
-import net.dv8tion.jda.api.requests.ErrorResponse
 import net.dv8tion.jda.api.utils.messages.MessageCreateData
-import net.dv8tion.jda.api.utils.messages.MessageEditData
-import java.util.concurrent.TimeUnit
 
 @CommandMarker
-class TextDocs(private val docIndexMap: DocIndexMap, private val components: Components) : TextCommand() {
+class TextDocs(
+    private val docIndexMap: DocIndexMap,
+    private val commonDocsController: CommonDocsController,
+    private val textDocsController: TextDocsController
+) : TextCommand() {
     @JDATextCommand(name = "docs", description = "Shows the documentation for a class, a method or a field")
     fun onTextDocs(
         event: BaseCommandEvent,
@@ -36,7 +34,7 @@ class TextDocs(private val docIndexMap: DocIndexMap, private val components: Com
             }
             else -> {
                 docIndex.getClassDoc(query)?.let {
-                    event.channel.sendMessage(CommonDocsHandlers.getDocMessageData(event.member, false, it, components)).queue()
+                    event.channel.sendMessage(commonDocsController.getDocMessageData(event.member, false, it)).queue()
                     return
                 }
 
@@ -44,39 +42,8 @@ class TextDocs(private val docIndexMap: DocIndexMap, private val components: Com
             }
         }
 
-        getDocSuggestionMenu(docIndex, suggestions)
+        textDocsController.getDocSuggestionMenu(docIndex, suggestions)
             .let { MessageCreateData.fromEditData(it.get()) }
             .also { event.channel.sendMessage(it).queue() }
     }
-
-    private fun getDocSuggestionMenu(docIndex: DocIndex, suggestions: List<DocSuggestion>) =
-        CommonDocsHandlers.buildDocSuggestionsMenu(docIndex, suggestions, components) {
-            useDeleteButton(true)
-
-            setTimeout(2, TimeUnit.MINUTES) { menu, message ->
-                menu.cleanup()
-                message!!
-                    .editMessageComponents()
-                    .queue(
-                        null,
-                        ErrorHandler().ignore(ErrorResponse.UNKNOWN_MESSAGE, ErrorResponse.UNKNOWN_WEBHOOK)
-                    )
-            }
-
-            setCallback { buttonEvent, entry ->
-                val identifier = entry.identifier
-                val doc = when {
-                    '(' in identifier -> docIndex.getMethodDoc(identifier)
-                    '#' in identifier -> docIndex.getFieldDoc(identifier)
-                    else -> docIndex.getClassDoc(identifier)
-                }
-
-                when (doc) {
-                    null -> buttonEvent.reply_("This item is now invalid, try again", ephemeral = true).queue()
-                    else -> buttonEvent.editMessage(MessageEditData.fromCreateData(
-                        CommonDocsHandlers.getDocMessageData(buttonEvent.member!!, false, doc, components)
-                    )).queue()
-                }
-            }
-        }
 }
